Do not require `from_raw(into_raw(v))` to be the same as `v` for safety
This is important for structs, since if we require that we must hide the underlying raw value since changing it can cause UB without `unsafe` code. If we don't require that, however, we can not hide it with a warning that changing it to an out-of-range value may have unexpected consequences.